- Fixed a major bug (not triggered, though)  - Added Inventory class.  - More functions!  - Things!
git-svn-id: https://subversion.assembla.com/svn/mapler-me@147 7cbc8f23-9b70-4854-b091-be65f742de0c

+++ branches/SomeFunnyDLL/MESession.cs
@@ -184,7 +184,7 @@ namespace System
                 }
                 else
                 {
-                    StartReceive((ushort)(_receiveLength - _receivePosition), true);
+                    StartReceive(_receiveLength - _receivePosition, true);
                 }
             }
             catch (Exception ex)
